Replace arrow icons (up, down, left, right
